Discord-Bot/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
discord.py==2.0.1
PyNaCl==1.5.0
python-dotenv==0.21.0
virtualenv==20.16.7
aiohttp-jinja2==1.5.0
aiohttp_session==2.9.0
aiohttp-security==0.4.0
aiohttp-apispec==2.2.3
jinja2==3.0.3
pyyaml>=5.1
cryptography>=3.2,<37.0.0; python_version <= '3.7'
cryptography>=3.2; python_version > '3.7'
websockets>=10.3
Sphinx==5.1.1
docutils==0.16 # Broken bullet lists in sphinx_rtd_theme https://github.com/readthedocs/sphinx_rtd_theme/issues/1115
sphinx_rtd_theme==0.4.3
myst-parser==0.18.0
marshmallow==3.5.1
dirhash==0.2.0
docker==4.2.0
donut-shellcode==0.9.2
marshmallow-enum==1.5.1
ldap3==2.8.1
dnspython==2.1.0
asyncssh==2.11.0
aioftp~=0.20.0; python_version >= '3.7'
aioftp==0.16.1; python_version < '3.7'
peewee==3.16.1